ivshmem-net: Fix bogus transition to RESET state
If we are in READY but the remote is still in INIT, we so far fell back
to RESET which caused the setup to get stuck. Fix this by only
transitioning from READY/RUN to RESET in ivshm_net_state_change if the
remote is in RESET as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
(cherry picked from commit db15735cea01865acbf2bad64333adb349d4d49d nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>