staging: fsl_ppfe/eth: add fixed-link support
In cases where MAC is not connected to a normal MDIO-managed PHY
device, and instead to a switch, it is configured as a "fixed-link".
Code to handle this scenario is added here.

phy_node in the dtb is checked to identify a fixed-link.
On identification of a fixed-link, it is registered and connected.

Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>
(cherry picked from commit df7df3175717761e42cbefefb4138c5e37b772d7 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>