clk: imx8m: make a check the m4 is enabled or not
Make a check that the M4 is enabled or not.
Otherwise, the clocks of the M4 maybe turned off later.

The UART clock used by M4 maybe turned off by Linux side, after the
initialization of the clocks. Enable the UART2 clock when M4 is enabled.

Squashed: clk: imx7d: enable uart2 clock when m4 is enabled

Signed-off-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>