ahci: qoriq: workaround for errata A-379364 on lx2160a
There is a erratum on lx2160a which is: "SATA link is
going down sometime during sata initialization"
The workaround for it is to reset the lane. This patch
implements this workaround.
This erratum only exists on lx2160 Rev1, will be addressed
on Rev2 and later.

Signed-off-by: Peng Ma <peng.ma@nxp.com>
(cherry picked from commit b624fc6d0f7e65ea095d53a55fb6765ee701ce9b nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>