LF-202-3 clk: imx: scu: ignore clks not owned
Not register clks that not owned to current partition.

Squashed: LF-279 clk: imx: scu: ignore cpu resources when do owned check

CPU resources are specical resources, it is assigned in ATF, not
non-secure OS, but we still need to allow cpu freq, so return
true for non-secure OS for cpu resources.

Signed-off-by: Peng Fan <peng.fan@nxp.com>
Tested-by: Anson Huang <anson.huang@nxp.com>
Reviewed-by: Leonard Crestez <leonard.crestez@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>