sdk_dpaa: ceetm: fix recursive dependencies
Due to dependencies between the fsl_ceetm and fsl_dpa modules, remove
the module support for the ceetm driver and integrate it into the main
DPAA Ethernet driver.

The registration of the CEETM Qdisc was the only operation done at
module init. Pass the management of the Qdisc register and unregister
operations to the loading and unloading of the DPAA driver.

Signed-off-by: Camelia Groza <camelia.groza@nxp.com>
(cherry picked from commit b83044c0eaabec9007a1e7a9625c0fa380478f1e nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>