LF-733: mtd: rawnand: gpmi: remove the BBM option in gpmi controller
When we tested the NAND on kernel 5.4, found bad block scan was skipped
since BBM options was not correctly set, as a workaround, we set it in
gpmi controller code. But this change is arbitrary since it doesn't know
which nand chip mounted.

The root cause was found and fixed by the following commit, which fixed
the issue in micron nand chip driver that deal with BBM for 2K page NAND
only, and change the nand_bbm_get_next_page not just skip the scan bad
block process when no BBM option found.

commit a3c4c2339f8948b0f578e938970303a7372e60c0
Author: Piotr Sroka <piotrs@cadence.com>
Date: Tue Sep 24 06:54:31 2019 +0100

With this fix, the previous BBM change in gpmi driver can be reverted.

Signed-off-by: Han Xu <han.xu@nxp.com>
(cherry picked from commit e7c052045afeeded8052ec0d7bb274daddbf5221 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>