MLK-23754-35 ARM: dts: imx7ulp: add cm4 node
Add cm4 node for remoteproc usage.

81fb06cb6bd4 ("remoteproc: Fall back to using parent memory pool if no dedicated available")
breaks i.MX7ULP SD Boot.

i.MX7ULP not use vdevbuffer for communication between Acore and Mcore.
However we still use memory-region for vring. Because we added
"shared-dma-pool" to the resered region, so when vdevbuffer not found,
the memory-region will be used for the virtio data buffer.

no need "shared-dma-buffer" to avoid the area to be used to data buffer,
then it will still use global dma area for data buffer.

Reviewed-by: Richard Zhu <hongxing.zhu@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>