LF-1373-7: drm/bridge: nwl-dsi: Add support for 8qm & 8qxp
The current nwl-dsi driver was initially written for 8mq, but the same
IP can also be found in 8qm and 8qxp SoCs.
The changes needed for 8qm and 8qxp are regarding to the architecture
differences:
- on 8qm and 8qxp, DSI reset can be accessed only through system
controller.
- also, on 8qm and 8qxp there are two nwl-dsi IP blocks, instead on just
  one, as found in 8mq
- on 8qxp, the dphy is a shared phy with the lvds IP block, so we need
  to account this too, when controlling the mipi dphy.

This patch addresses the above and fixes different problems that occured
with the addition of these new features.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20210514 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20220315 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
[ Liu Ying: Fix conflicts to upgrade to next-20220422 ]
Signed-off-by: Liu Ying <victor.liu@nxp.com>
(cherry picked from commit 09bc7b39b05916be0bd91ed5f2012506afb4b242 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>