MLK-24805-1: drm/panel: rm67191: Add support for rm67199
There is a new and improved version of this panel, which is called
rm67199, and is the same as rm67191 (in size, aspect and resolution).
Since, only the DSI initialization sequence differs, I added this
into the existing driver for rm67191.
So:
- compacted the DSI commands into 4 per line so they will not consume
too many lines of code
- updated the prepare sequence, since the instructions for this panel
  were provided (the sleeps for rm67191 were just guesses)
- added platform data to determine which version of the panel is used
  with this driver

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@nxp.com>
(forward ported from commit 0bb5b57ebec6b4b2d668951a358e2ee2063d8716 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>