ivshmem-net: Fix and rework carrier management
The carrier of our virtual link is supposed to go on only when both
sides are in RUN state. So far, it only went on when one side entered
RUN by being int READY before and not when RUN was reached via
ivshm_net_open().

Fix this by moving the carrier management into ivshm_net_run and
ivshm_net_do_stop, respectively. Make sure that netif_carrier_on is only
called when the peer in in RUN state already.

Reported-by: Jiafei Pan <Jiafei.Pan@nxp.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>