LF-3016-3 tools/virtio: ivshmem-console: correct device_vector to 0
We need to set device_vector to 0, otherwise inmate is not able to
trigger interrupt to root cell and virtio console not work.

This is a hack, not a good fix. With MSIX, we might able to use 2,
but we are using INTX now.

Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
(cherry picked from commit f0b1f844b6b4769e0b795194e4dcf920c0a3280e nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>