arm64: dts: freescale: Switch from clock-indices to bit-offsets
This is a temporary fix after rebasing the linux-nxp/dts/imx8
on top of next-20210514. The reason we're doing the switch is to
keep the dts working. The upstream version of the dts files already
use clock-indices. The NXP's tree needs to become ready to do so
too.

Signed-off-by: Abel Vesa <abel.vesa@nxp.com>
(cherry picked from commit 7de02d4682e5d3e3542fa252afb8556fb79f37c3 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>