LF-4836 arm64: imx8mx vpu: make sure output streamon cmd is seet to daemon after buffer is queued
the cmd V4L2_DAEMON_VIDIOC_STREAMON_OUTPUT should be sent to daemon
when the minimum needed buffers are queued.
so we should make sure that when buffers are queued, we send it to daemon.
These is a bug that driver may not send it to daemon due to some statue
issue.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Shijie Qin <shijie.qin@nxp.com>
(forward ported from commit 7792c31ad9b6223d46512464fbbd202b7f534f11 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>