MLK-20697-1: drm/panel: panel-raydium-rm67191: Add MIPI_DSI_MODE_NO_EOT_PACKET flag
This flag was automatically used in the DSI driver, making this panel
work, but the way that flag was used was wrong. Since, fixing the use of
this flag in the DSI driver now breaks this panel, just add it here.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
Reviewed-by: Liu Ying <victor.liu@nxp.com>
(cherry picked from commit c8dced644bc4c62faa272e71d98cc087ffc4a880 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>