LF-6139: arm64: dts: fix the iMX8X cm4 node build warning
The new implementation of_get_cpu_hwid() in imx_rproc.c parsing the
"reg" in cm4 node, but "reg" keywords refer to reg_format and in this
node it conflicts with address-cells/size-cells rule and causes build
warning. Put the node in a cm4 node and set the correct
address-cells/size-cells value to fix the issue.

Signed-off-by: Han Xu <han.xu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>