dpaa2-switch: say that the bond device is offloaded as a bridge port
We have to let the switchdev framework know if a specific bridge port is
offloaded or not, even if that port is bond device.
For this to happen, create the dpaa2_switch_port_to_bridge_port function
which will determine the bridge port corresponding to a particulat dpaa2
switch interface and use it in the switchdev_bridge_port_offload call.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(forward ported from commit 53cb33661a41e1c15e46280b7459156d0bcf8019 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>