LF-6561: arm64: dts: imx8mpevk: Enable ALDPS mode of the FEC PHY
The Advance Link Down Power Saving mode of the PHY of the FEC was
enabled in the version 5.10 on i.MX8MEK-EVK board. But in the version
5.15, the ALDPS mode was disabled, which results in higher power
consumption.
Checked with the last committer, the ALDPS mode of the PHY of the
FEC can be enabled, but need to confirm that there is no problem with
the suspend & resume test. After the suspend & resume testing with
the ALDPS mode enabled, the result of the test is passed. Therefore,
this patch add the "realtek,aldps-enable" property in the relevant
phy node to enable ALDPS mode.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>