arm64: dts: ls1028a-rdb-dpdk: remove pause property
DPDK is not capable to handle any type of congestion
or back-pressure. So if there is congestion on the network,
HW can generate PAUSE frames to limit the traffic and packets
can be dropped in the DPDK.

So in this patch, removing the pause property from switch and
ethernet ports.

Signed-off-by: Gagandeep Singh <g.singh@nxp.com>
(cherry picked from commit 97eca6701670d664ba26035dedd287aca7031c4c nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>