arm64: lsdk.config: make tc-taprio built-in
Since upstream commit 10ed11ab6399 ("net: dsa: felix: build as module
when tc-taprio is module"), CONFIG_NET_DSA_MSCC_FELIX depends on
NET_SCH_TAPRIO, which in the arm64 defconfig is a module.

This causes the Felix DSA switch driver to now be built as module, which
is inconvenient in some setups.

Make CONFIG_NET_SCH_TAPRIO built-in to allow the Felix driver to be
built-in again.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>