ion: Add ION UAPI headers
On kernel versions >= 5.15, ION has been completely deprecated, meaning
that the kernel no longer exposes the ION UAPI. There are versions of
libion on operating systems (e.g. Linux Embedded (LE)) that rely on the
kernel to export the latest ION headers to build properly.
Having LE use the same strategy as Android's libion (i.e. maintaining a
copy of the new UAPI and using that to supplement the legacy linux/ion.h
with the new UAPI definitions) is not feasible. It is not feasible because
LE's libion includes the msm-kernel headers, and certain kernels export
the non-legacy ION UAPI (kernels >= 4.14) as linux/ion.h, which causes
libion to use the kernel header for linux/ion.h instead of the legacy
linux/ion.h. When including a local copy of the updated ION UAPI in the
LE libion project, redefinition errors are observed when building libion
with the kernel headers from a kernel that supports the non-legacy
version of ION, as the local copy will redefine the same structures.
Android's libion does not have this problem, as it doesn't use the
msm-kernel's headers, but only its own, so its linux/ion.h is always the
legacy linux/ion.h.
It is possible to stop including the msm-kernel headers for libion,
which would make using Android's approach feasible. However, this
introduces issues when libdmabufheap tries to include ion_4.12.h for the
new structure definitions of the heap query ioctl, as these definitions
redefine the ones obtained from the msm-kernel's linux/ion.h when
building libdmabufheap. Libdmabufheap must be built with the
msm-kernel's UAPI headers, as it relies on the dma-heap.h UAPI file.
Thus, add the ION UAPI headers as part of the headers that the kernel
will expose as of android12-5.10 commit b4604acd52a6
("UPSTREAM: USB: gadget: detect too-big endpoint 0 requests").
This change should not be carried forward to kernels that will run on LE
versions that do not need libion.

Change-Id: I53831911a9aeebbc70edf421e984bd3b1e2b9c36
Signed-off-by: Isaac J. Manjarres <quic_isaacm@quicinc.com>
Signed-off-by: Anurag Chouhan <quic_achouhan@quicinc.com>
(cherry picked from commit f5f7b5699e2a9cc6954619e59f64cb209e4834a5)
Signed-off-by: Nathan Sweetman <nathan.sweetman@canonical.com>