LF-7511-1 input: touchscreen: exc300: read the second frame during the interrupt line keep low
For the EETI EXC series, it support up to 10 points, two consecutive frames
need to read. But for each frame, need to be read when the interrupt line
keep in low level, otherwise, the read frame always get all 0xff data.

The hardware behavior for this touch device is it only support the low level
trigger method, when irq line change to low level, usr need to read the
first frame, and then the irq line return to high level, after about 320us,
the irq line change to low level, then usr need to read the second frame.

Without this patch, we always find the second frame is readed during the irq
line keep in high level, through the i2c logic analyzer, we see the data in
the second frame is all 0xFF. This is because usr read the second frame
immediately after the first frame in the same irq handler. But the irq line
will back to high level after the first frame and keep about 320us. This
behavior is wrong, and even worse, in some probability, this wrong behavior
will cause the touch device stuck, can't generate interrupt any more, which
means the interrupt line always keep high even after touch the screen.

This patch change to read only one frame in one irq handler. For the case
need to read two frames, in the first irq handler, save the total contact
number, then in the second irq handler, report the leftover contact.

Fixes: 7e577a17f2ee ("Input: add I2C attached EETI EXC3000 multi touch driver")
Reviewed-by: Carlos Song <carlos.song@nxp.com>
Signed-off-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>