LF-7815 remoteproc: imx_rproc: support starting Cortex-M33 from non-TCM address for i.MX93
The current implementation not pass the M33 firmware loading address to
Arm Trusted Firmware(ATF), so that implies the M33 firmware is using
address 0xFFE0000. If the M33 firmware using a new address, it will not
boot. So parsing the elf file to get the address of section ".interrupts"
which is the first section of the elf file as bootaddr and pass to ATF.

Reviewed-by: Ye Li <ye.li@nxp.com>
Signed-off-by: Peng Fan <peng.fan@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>