LF-7833: ARM: config: Add TUN device support
Since the project Matter needs the support of TUN device to work
as wpan0 virtual interface, so enable CONFIG_TUN for arm32 devices.
Also, the TUN is a universal network tunnel driver and might be used
by other user space programs in the future. It's fine to enable
CONFIG_TUN by default.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Jun Li <jun.li@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>