LF-7981-8: drm: bridge: it6161: Use special horizontal flags
This IP does not work when the horizontal blaking area is in pixels, but
works when this area is in bytes.
Add the flags for disabling the horizontal blanking area, so that the
DSI host can take care of this and send the blankings in bytes instead
of pixels.

Signed-off-by: Robert Chiras <robert.chiras@nxp.com>
(cherry picked from commit 40fe2c7dff08a3e59d5c309d368b37f493207a78 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>