LF-8234 arm: imx: Remove LDO 1P0D disable in low power idle
the LDO 1P0D is used for the digital portion of the MIPI and
PCIE modules. on 6.1 and later,The PCIE will never probe failure
even if no PCIE EP device connected. If the 1P0D regulator is
disable/enable when system enters low power idle, the PCIE logic
will be in unknown status, when doing PCIE suspend, it will lead
to system hang.

1P0D disable/enable should be removed from low power idle flow
to make sure PCIE can work as we expected.

The reason why we don't have such issue on previous release is that
if no PCIE card connected, PCIE will directly probe fail and the
PHY link will be down, no PCIE driver suspend/resume when system
enter suspend.

Signed-off-by: Jacky Bai <ping.bai@nxp.com>
Reviewed-by: Peng Fan <peng.fan@nxp.com>
Acked-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>