LF-8475 arm64: dts: imx8mq-wevk: enable the front M.2 connector on imx8mq-wevk board
The main change for imx8mq-wevk board is to deprecate the on-board
cypress wifi and add a new M.2 on the front of the board. This new
imx8mq-wevk.dts enables the front M.2 connector for WiFi and Bluetooth.

Signed-off-by: Sherry Sun <sherry.sun@nxp.com>
(forward ported from commit f4ff1a6cb7e6ff019c4bf0ef6206121644445493 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>