LF-4131 iio: gyro: fxas21002c: Fix raw data is not updated in trigger/buffer
The runtime PM autosuspend is used for the sensor. When the sensor
in suspend runtime PM status, it is in standby mode and will not update
raw data. When read raw data into buffer by trigger, the data which is
read into the buffer in each trigger interrupt is constant.

Resume the sensor before read new data into the buffer by trigger and
put the sensor into autosuspend runtime PM status after that.

Signed-off-by: Carlos Song <carlos.song@nxp.com>
Reviewed-by: Haibo Chen <haibo.chen@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>