phy: lynx-28g: add support for 25GBASER
Add support for 25GBASE-R in the Lynx 28G SerDes PHY driver.
This mainly means being able to determine if a PLL is able to support
the new interface type, to determine at probe time if a lane is
configured from the Reset Configuration Word (RCW) with this interface
type and to be able to reconfigure a lane.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(cherry picked from commit de02207736a43edfbc69752da06546ad7cbf893c nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>