arch: arm64: lx2160a-rdb: describe the retimer as a Generic PHY
The LX2160ARDB board has a Inphi Retimer PHY which is used on the 2
SerDes which go to the SFPs. Model this Retimer PHY as a generic PHY and
extend the MAC DT nodes to also reference the retimer.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(cherry picked from commit 235e302decb63947be4a0c30555077866d8d6c77 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>