LF-8868: media: ov5640: fix incorrect frame frate issue
If run OV5640 with 640x480@30 setting after power up, the real
frame rate for it is 60, not 30. The reason is default frame
interval parameter in probe is 30 but default link frequency is
to generate 60 frame rate, so correct it.

Signed-off-by: Guoniu.zhou <guoniu.zhou@nxp.com>
Reviewed-by: Robby Cai <robby.cai@nxp.com>
(cherry picked from commit dd40f93693a927e717a61a5624490c208765f2ed nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>