LF-9573: arm: imx: keep bit 10 (1T/2T mode) during frequency change
This patch is to set the 10th bit in the DDRC_MSTR register to be the
same as the existing value. This ensures that the 1T/2T mode is preserved
after the frequency is changed.

Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Reviewed-by: Jacky Bai <ping.bai@nxp.com>
(cherry picked from commit 4a6a7a6f62eca69bc20418210b5551fd49b07aec nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>