arm64: dts: ls1088ardb: update all ethernet ports with the SerDes phy info
Update all ethernet ports so that they have a link to the SerDes phy
associated.

Also, minimally describe the SFP cage featured on the board.  Please
note that it seems link the mod-def0 GPIO used for detecting if a SFP
module is inserted or not is not exported as a value in the QIXIS CPLD
on this board. Because of this, if the SFP module is removed while Linux
is running on the board, the sfp subsystem will print out errors since
it tries to access an SFP which is not connected. This is true only when
the port is configured with MAC_LINK_TYPE_BACKPLANE or
MAC_LINK_TYPE_PHY.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>
(cherry picked from commit 9be481604785e34710747515e86f03bd95d12107 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>