phy: lynx-28g: don't concatenate macros for lynx_28g_lane_rmw() args "val" and "mask"
The fact that lynx_28g_lane_rmw() prepends "val" with LYNX_28G_${reg}_
means that we can't specify numerical values or other expressions for
this argument.

Refactor the code to fully spell out the string concatenation. This has
the added readability benefit that searching these arguments gets us
directly to their definition in the code.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>