LF-9720: arm64:dts:freescale:imx8mn: Fix typo on imx8mn-pinfunc.h for SAI3_RXD_PDM_BIT_STREAM1
imx8mn-pinfunc.h had an incorrect input_val for SAI3_RXD_PDM_BIT_STREAM1

checkpatch reported the following error and warning:

WARNING: line length of 104 exceeds 100 columns
+#define MX8MN_IOMUXC_SAI3_RXD_PDM_BIT_STREAM1 0x1D4 0x43C 0x538 0x6 0xA

ERROR: Macros with complex values should be enclosed in parentheses
+#define MX8MN_IOMUXC_SAI3_RXD_PDM_BIT_STREAM1 0x1D4 0x43C 0x538 0x6 0xA

total: 1 errors, 1 warnings, 8 lines checked

For the error, due to the pinctrl code design, we can't reslove the
error as the checkpatch told that will result in dts parse error.
So, we will not fix it and keep the patch as is.(the same as before)
For the warning, for better coding reading, we will not fix it too.

Reviewed-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Gustavo Arambula <gustavo.arambula@nxp.com>
[update the commit history to add checkpatch information]
Signed-off-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>