LF-9861: dmaengine: imx-sdma: Fix audio p2p test on imx8m
The sdma p2p script presently does not deal with the transactions involving
two devices connected to AIPS bus.

So the audio p2p test can use spba-bus, but the sdma driver check the
peripherals address doesn't belong to the spba-bus due to get wrong
spba-bus address when there are multi spba-bus. The sdma controllers
and the peripherals are fixed to the correspond spba-bus.

This patch modifies the method of finding spba-bus, and make sure the
parent node of the spba-bus is same with the sdma.

Reported-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Reviewed-by: Shengjiu Wang <shengjiu.wang@nxp.com>
Signed-off-by: Joy Zou <joy.zou@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>