net: ocelot: keep force forward port in forward mask
The 802.1CB implementation from Felix uses FDB entries to identify the
streams. If the stream is forwarded to a port that is not in the
switch's forwarding mask, the 802.1CB module will not work.

This patch keep the ports with forced forwarding in the forwarding mask
of the switch whenever the port status changes.

Signed-off-by: Xiaoliang Yang <xiaoliang.yang_1@nxp.com>
(cherry picked from commit 52d8bd8f9d47ac5944facc4b37c3fc70c5315fe3 nxp-kernel)
Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>