LF-10640-6 ptp: add NETC PTP driver support for i.MX95
The i.MX95 NETC's Timer is different from the LS1028A NETC's Timer. In
addition to swapping the offsets of the upper 32-bit register and the
lower 32-bit register and adding ALRAM and FIPER control registers,
there are some main registers (such as the TMR_CTRL and TMR_ADD) that
are used in different ways even though they have the same functionality.

Yes, we can reuse the PTP driver of LS1028A and make some modifications
to support i.MX95 NETC's Timer, but this will undoubtedly increase the
complexity of the code and make the driver difficult to maintain. So we
have added a new driver named ptp_netc to support NETC Timer of i.MX95
and future i.MX SoCs such as i.MX943, i.MX952, etc.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Acked-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>