MLK-26147: net: stmmac: dwmac-imx: request high frequency mode
Some i.MX SoCs like the i.mx8mq support adjusting the frequency of the
DDR, AHB, and AXI buses based on system loading. If the dwmac interface
in the driver does not request a HIGH frequency, it can significantly
degrade performance when the system switches to a lower frequency to
conserve power.

For example, on an i.MX8MQ EVK board, the throughput dropped to around
100Mbit/s on a 1Gbit connection:

    [ ID] Interval           Transfer     Bitrate
    [  5]   0.00-10.00  sec   117 MBytes  97.9 Mbits/sec

However, throughput can return to expected levels after its driver requests
the high frequency mode. Requesting high frequency in the dwmac driver is
essential to maintain full throughput when the i.MX SoC adjusts bus speeds
for power savings.

Signed-off-by: Mario Castaneda <mario.ignacio.castaneda.lopez@nxp.com>
Signed-off-by: Shenwei Wang <shenwei.wang@nxp.com>
Tested-by: Mario Castaneda <mario.ignacio.castaneda.lopez@nxp.com>
Signed-off-by: Ondrej Kubik <ondrej.kubik@canonical.com>